requirements.txt
torch==1.10.2+cu113
numpy==1.22.1
opencv-python==4.5.5.64
python-mnist==0.7